AMB-961 rename artifact name and base namespace to converter
