+++ docker/Transformation-Engine/src/main/java/nhsd/fhir/converter/FhirConverterApplication.java
@@ -1,4 +1,4 @@
-package nhsd.fhir.transformationenginepoc;
+package nhsd.fhir.converter;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
