+++ docker/Transformation-Engine/src/main/java/nhsd/fhir/converter/controller/StatusController.java
@@ -1,4 +1,4 @@
-package nhsd.fhir.transformationenginepoc.controller;
+package nhsd.fhir.converter.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
