+++ docker/Transformation-Engine/src/main/java/nhsd/fhir/converter/service/converter/Converter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nhsd.fhir.transformationenginepoc.service.converter;
+package nhsd.fhir.converter.service.converter;
 
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.context.FhirVersionEnum;
