+++ docker/Transformation-Engine/src/test/java/nhsd/fhir/converter/FhirConverterApplicationTests.java
@@ -1,4 +1,4 @@
-package nhsd.fhir.transformationenginepoc;
+package nhsd.fhir.converter;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
