+++ docker/Transformation-Engine/src/test/java/nhsd/fhir/converter/controller/ConversionControllerUnitTest.java
@@ -1,6 +1,6 @@
-package nhsd.fhir.transformationenginepoc.controller;
+package nhsd.fhir.converter.controller;
 
-import nhsd.fhir.transformationenginepoc.service.ConversionService;
+import nhsd.fhir.converter.service.ConversionService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
