+++ docker/Transformation-Engine/src/test/java/nhsd/fhir/converter/controller/StatusControllerTest.java
@@ -1,4 +1,4 @@
-package nhsd.fhir.transformationenginepoc.controller;
+package nhsd.fhir.converter.controller;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
