+++ docker/Transformation-Engine/src/test/java/nhsd/fhir/converter/service/ConversionServiceTest.java
@@ -1,4 +1,4 @@
-package nhsd.fhir.transformationenginepoc.service;
+package nhsd.fhir.converter.service;
 
 import org.apache.commons.io.FileUtils;
 import org.json.JSONObject;
