+++ docker/Transformation-Engine/src/test/java/nhsd/fhir/converter/service/ConverterTest.java
@@ -1,9 +1,9 @@
-package nhsd.fhir.transformationenginepoc.service;
+package nhsd.fhir.converter.service;
 
 import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.context.FhirVersionEnum;
 import ca.uhn.fhir.parser.IParser;
-import nhsd.fhir.transformationenginepoc.service.converter.Converter;
+import nhsd.fhir.converter.service.converter.Converter;
 import org.hl7.fhir.convertors.advisors.impl.BaseAdvisor_30_40;
 import org.hl7.fhir.convertors.conv30_40.VersionConvertor_30_40;
 import org.hl7.fhir.dstu3.model.Patient;
