+++ docker/Transformation-Engine/src/test/java/nhsd/fhir/converter/service/transformers/MedicationRequestConverterTest.java
@@ -1,7 +1,7 @@
-package nhsd.fhir.transformationenginepoc.service.transformers;
+package nhsd.fhir.converter.service.transformers;
 
 import ca.uhn.fhir.context.FhirVersionEnum;
-import nhsd.fhir.transformationenginepoc.service.converter.MedicationRequestConverter;
+import nhsd.fhir.converter.service.converter.MedicationRequestConverter;
 import org.apache.commons.io.FileUtils;
 import org.json.JSONException;
 import org.json.JSONObject;
